% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\docType{data}
\name{GAMS_EXECUTION_ERROR}
\alias{GAMS_EXECUTION_ERROR}
\title{Error code returned by GAMS in case of an execution error}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
GAMS_EXECUTION_ERROR
}
\description{
Error code returned by GAMS in case of an execution error
}
\keyword{datasets}
