% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\docType{data}
\name{GAMS_FILE_ERROR}
\alias{GAMS_FILE_ERROR}
\title{Error code returned by GAMS in case of a file error}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
GAMS_FILE_ERROR
}
\description{
Error code returned by GAMS in case of a file error
}
\keyword{datasets}
