% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\docType{data}
\name{GAMS_NORMAL_RETURN}
\alias{GAMS_NORMAL_RETURN}
\title{GAMS return code in case of normal completion}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
GAMS_NORMAL_RETURN
}
\description{
GAMS return code in case of normal completion
}
\keyword{datasets}
