% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\docType{data}
\name{GAMS_PARAMETER_ERROR}
\alias{GAMS_PARAMETER_ERROR}
\title{Error code returned by GAMS in case of a faulty parameter}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
GAMS_PARAMETER_ERROR
}
\description{
Error code returned by GAMS in case of a faulty parameter
}
\keyword{datasets}
