% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdx.R
\name{gdx_info}
\alias{gdx_info}
\title{Return info on content of a GDX}
\usage{
gdx_info(dir, name = GDX_FILE_NAME)
}
\arguments{
\item{dir}{Directory containing GDX file.}

\item{name}{Name of GDX file.}
}
\value{
List describing GDX content.
}
\description{
Returns a list with information on the content of a GDX file. Uses
\code{GDX_FILE_NAME} by default, which is the name used by \code{run()} to dump
an end-of-execution GDX. When using this default, the \code{dir} argument
should be set to the redirection directory used for the \code{run()}.
}
