% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdx.R
\name{gdx_parameter}
\alias{gdx_parameter}
\title{Read a parameter from a GDX file as a tibble}
\usage{
gdx_parameter(info, name)
}
\arguments{
\item{info}{Info list as returned by \code{\link[=gdx_info]{gdx_info()}}.}

\item{name}{Name of the parameter to read.}
}
\value{
Tibble with parameter content.
}
\description{
Read a parameter from a GDX file as a tibble
}
