% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdx.R
\name{gdx_set}
\alias{gdx_set}
\title{Read a set from a GDX file}
\usage{
gdx_set(info, name)
}
\arguments{
\item{info}{Info list as returned by \code{\link[=gdx_info]{gdx_info()}}.}

\item{name}{Name of the set to read.}
}
\value{
List with set content.
}
\description{
Read a set from a GDX file
}
