% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_dir.R
\name{get_sys_dir}
\alias{get_sys_dir}
\title{Get GAMS system directory}
\usage{
get_sys_dir()
}
\value{
Path to the in-use GAMS system directory.
}
\description{
Get the path to the GAMS system directory being used for testing. This
is the GAMS installation directory where the GDX libraries and GAMS
compiler/interpreter will be loaded from. It therefore determines which
GAMS version is used for testing. The GAMS system directory can be
set with \code{set_sys_dir()}.
}
\details{
When no system directory is set yet, this function will try to get and
set a default system directory from an environment variable, first
\code{R_GAMS_SYSDIR}, and if not available via \code{PATH} (on Windows) or
\code{LD_LIBRARY_PATH} (on Linux). If no such default is available either,
this function throws an error.
}
\examples{
sys_dir <- get_sys_dir()
}
