% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_dir.R
\name{local_re_dir}
\alias{local_re_dir}
\title{Create temporary redirection directory with local cleanup}
\usage{
local_re_dir(env = parent.frame())
}
\arguments{
\item{env}{[\code{environment}: \code{parent.frame()}]\cr Attach cleanup handler to
this environment, defaults to the calling environment.}
}
\value{
re_dir Newly created uniquely named temporary directory.
}
\description{
Create a temporary directory via \code{create_re_dir()} with a cleanup handler
attached to the calling environment. Cleanup deletes the directory and its
contents using \code{withr::defer()}.
}
\examples{
re_dir <- local_re_dir()
# Do something with re_dir.
# Cleanup happens when this environment exits/returns
}
