% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\docType{data}
\name{PAR_FILE_NAME}
\alias{PAR_FILE_NAME}
\title{File name of GAMS parameter file used for a \code{\link[=run]{run()}} and written to \code{re_dir}}
\format{
An object of class \code{character} of length 1.
}
\usage{
PAR_FILE_NAME
}
\description{
File name of GAMS parameter file used for a \code{\link[=run]{run()}} and written to \code{re_dir}
}
\keyword{datasets}
