% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_lines}
\alias{read_lines}
\title{Read line from a text file}
\usage{
read_lines(path)
}
\arguments{
\item{path}{Path to text file.}
}
\value{
Character vector holding lines in the read text file.
}
\description{
Read line from a text file
}
