% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_log}
\alias{read_log}
\title{Read lines from log file}
\usage{
read_log(re_dir)
}
\arguments{
\item{re_dir}{Redirection directory used for the \code{\link[=run]{run()}}.}
}
\value{
Character vector holding lines in the read log file.
}
\description{
Reads the lines from the GAMS log output file written during a \code{\link[=run]{run()}}.
}
