% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_lst}
\alias{read_lst}
\title{Read lines from listing file}
\usage{
read_lst(dir, name = LST_FILE_NAME)
}
\arguments{
\item{dir}{Directory containing log file.}

\item{name}{Name of log file.}
}
\value{
Character vector holding lines in the read listing file.
}
\description{
Reads the lines from a GAMS listing file. Uses\code{LST_FILE_NAME} by default,
which is where \code{run()} redirects GAMS listing to. When using this
default, the \code{dir} argument should be set to the redirection directory
used for the \code{run()}.
}
