% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_par}
\alias{read_par}
\title{Read lines from parameter file}
\usage{
read_par(re_dir)
}
\arguments{
\item{re_dir}{Redirection directory used for the \code{\link[=run]{run()}}.}
}
\value{
Character vector holding lines in the read parameter file.
}
\description{
Reads the lines of the GAMS parameter file used for a \code{\link[=run]{run()}}.
}
