% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_ref}
\alias{read_ref}
\title{Read lines from reference file}
\usage{
read_ref(re_dir)
}
\arguments{
\item{re_dir}{Redirection directory used for the \code{\link[=run]{run()}}.}
}
\value{
Character vector holding lines in the read reference file.
}
\description{
Reads the lines of the GAMS reference output file written during a \code{\link[=run]{run()}}.
}
