% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_trace}
\alias{read_trace}
\title{Read lines from trace file}
\usage{
read_trace(re_dir)
}
\arguments{
\item{re_dir}{Redirection directory used for the \code{\link[=run]{run()}}.}
}
\value{
Character vector holding lines in the read trace file.
}
\description{
Reads the lines from the GAMS trace file written during a \code{\link[=run]{run()}}.
}
