% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_trace_report}
\alias{read_trace_report}
\title{Read lines from trace report file}
\usage{
read_trace_report(re_dir)
}
\arguments{
\item{re_dir}{Redirection directory used for the \code{\link[=run]{run()}}.}
}
\value{
Character vector holding lines in the read trace report file.
}
\description{
Reads the lines of the GAMS trace report file written during \code{\link[=report_trace]{report_trace()}}.
}
