% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_trace_summary}
\alias{read_trace_summary}
\title{Read lines from trace summary file}
\usage{
read_trace_summary(re_dir)
}
\arguments{
\item{re_dir}{Redirection directory used for the \code{\link[=run]{run()}}.}
}
\value{
Character vector holding lines in the read trace summary file.
}
\description{
Reads the lines of the GAMS trace summary file written during \code{\link[=report_trace]{report_trace()}}.
}
