% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_dir.R
\name{set_sys_dir}
\alias{set_sys_dir}
\title{Set GAMS system directory}
\usage{
set_sys_dir(gams_sys_dir)
}
\arguments{
\item{gams_sys_dir}{Path to GAMS system directory.}
}
\description{
Set the GAMS installation to use for testing by specifying the system
directory where the GAMS binaries and libraries are installed. There
can be multiple GAMS versions installed, so this functions selects
the GAMS version to test with. Specifically, the GDX libraries and
GAMS compiler/interpreter located in the set GAMS system directory
will be used by the functions in the testGAMS package.
}
\details{
Internally, this function uses the \code{igdx()} function from the gdxrrw
package via which GDX files are accessed. The currently-set system
directory is returned by \code{get_sys_dir()}.

This function throws an error if the given directory cannot be set.
}
