% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\docType{data}
\name{TRACE_FILE_NAME}
\alias{TRACE_FILE_NAME}
\title{File name of GAMS trace file produced by \code{\link[=run]{run()}} in \code{re_dir}}
\format{
An object of class \code{character} of length 1.
}
\usage{
TRACE_FILE_NAME
}
\description{
File name of GAMS trace file produced by \code{\link[=run]{run()}} in \code{re_dir}
}
\keyword{datasets}
