% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\docType{data}
\name{TRACE_REPORT_FILE_NAME}
\alias{TRACE_REPORT_FILE_NAME}
\title{File name of GAMS trace report file produced by \code{trace_report()} in \code{re_dir}}
\format{
An object of class \code{character} of length 1.
}
\usage{
TRACE_REPORT_FILE_NAME
}
\description{
File name of GAMS trace report file produced by \code{trace_report()} in \code{re_dir}
}
\keyword{datasets}
