% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_lines}
\alias{write_lines}
\title{Write lines to a file}
\usage{
write_lines(path, lines)
}
\arguments{
\item{path}{Path of the file to create.}

\item{lines}{Character vector of lines to write.}
}
\description{
Truncates or creates the file before writing.
}
