% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_log}
\alias{read_log}
\title{Read lines from log file}
\usage{
read_log(dir, name = LOG_FILE_NAME)
}
\arguments{
\item{dir}{Directory containing log file.}

\item{name}{Name of log file.}
}
\value{
Character vector holding lines in the read log file.
}
\description{
Reads the lines from a GAMS log file. Uses\code{LOG_FILE_NAME} by default,
which is where \code{run()} to redirects GAMS logging to. When using this
default, the \code{dir} argument should be set to the redirection directory
used for the \code{run()}.
}
