% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_lst}
\alias{read_lst}
\title{Read lines from listing file}
\usage{
read_lst(re_dir)
}
\arguments{
\item{re_dir}{Redirection directory used for the \code{\link[=run]{run()}}.}
}
\value{
Character vector holding lines in the read listing file.
}
\description{
Reads the lines from the GAMS listing output file written during a \code{\link[=run]{run()}}.
}
