% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run a GAMS script for testing.}
\usage{
run(script, re_dir)
}
\arguments{
\item{script}{Path of GAMS script to run.}

\item{re_dir}{Redirection directory for holding GAMS output files.}
}
\value{
GAMS status/error/return code.
}
\description{
Run a GAMS script and captures output files for testing. The logging
output, listing file, trace file, and a GDX dump of all symbols are
redirected to files located in \code{re_dir} with names LOG_FILE_NAME,
LST_FILE_NAME, TRACE_FILE_NAME, and GDX_FILE_NAME respectively.
The default compile-and-execute action is used.
}
