% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_lines_cleanup}
\alias{write_lines_cleanup}
\title{Write lines to a file with deferred cleanup}
\usage{
write_lines_cleanup(path, lines, env = parent.frame())
}
\arguments{
\item{path}{Path of the file to create.}

\item{lines}{Character vector of lines to write.}

\item{env}{Environment to set \code{withr::defer()} cleanup on.}
}
\description{
Truncates or creates the file before writing, and sets deferred deletion
of the file to occur when the specified environment is destroyed. By
default, the invoking environment is set.
}
